Update adk_whitepaper.tex
Hello, I have edited the first page, up to line #156. 

Please review to make sure it is still technically accurate.

I do have a couple of questions: 

Line #70: Is coPow completely new and unique to Aidos Kuneen? Or is it built on an existing technology?

Line #149 and #151: Line #149 says that the transaction is added to iMesh immediately, Line #151 says that the bigger iMesh gets the faster the transaction confirms - these two lines seem to contradict each other.

Please let me know if you are happy with my work and would like me to keep going with the rest of the document.