Update adk_whitepaper.tex
Here is page 2, I have stopped at line #216 (to await your answer to the comments below).  I think I will only submit one page at time for you to look at otherwise the comments would be a lot to deal with, I will keep working ahead on an offline copy while you are reviewing this page.

Some comments:

#187 to #189 appears to be discussing potential options for providing anonymity to ADK, would this be more appropriate in the section above (under the heading "Strong Anonymity")? - I propose to move this section to the "Strong Anonymity" section so that under "Related Works" we talk about Monero and Bytecoin and then go straight to talking about Iota. - Please let me know if you agree and I will combine and re-write appropriately.

I wonder if we should mention that the code was originally forked from Iota - if we do not mention it people may accuse us of trying to hide it.  I have mentioned it in line #193 but it can be removed if required.

Lines #179 and #191 are intended to be subheadings - can you please make them bold in the final version

We need a source for line #203

Lines #203 to #215  can you please look at this extra closely to make sure it is technically accurate.